refactor: перенести исключения заявок в DomainTypes, разобрать InsufficientContactsException
ClaimAlreadyPresentException, OnlyOneApprovedClaimException,
ClaimTargetIsNotAcceptingClaims и InsufficientContactsException не зависят от
DataModel — это плоские JoinRpgBaseException, а база лежит в JoinRpg.Helpers,
на который DomainTypes уже ссылается. Переезжают в
DomainTypes/Characters/Claims/Exceptions.cs, чтобы правила валидации заявок
могли переехать в DomainTypes следом (ADR013).

ClaimWrongStatusException остаётся в JoinRpg.Domain: ему нужна EF-сущность Claim.

Заодно InsufficientContactsException добавлен в JoinMvcControllerBase:
раньше он проваливался в default-ветку, и игрок вместо объяснения, каких
контактов не хватает, видел "Неожиданная ошибка, обратитесь в техподдержку".
