+++ src/JoinRpg.Domain/Exceptions.cs
@@ -148,23 +148,6 @@ public class EntityWrongStatusException : JoinRpgProjectEntityException
     }
 }
 
-public class ClaimAlreadyPresentException : JoinRpgBaseException
-{
-    public ClaimAlreadyPresentException() : base("Claim already present for this character or group.") { }
-}
-
-public class OnlyOneApprovedClaimException : JoinRpgBaseException
-{
-    public OnlyOneApprovedClaimException() : base("Approved claim already present for this player, and project allows only one character.") { }
-}
-
-public class ClaimTargetIsNotAcceptingClaims : JoinRpgBaseException
-{
-    public ClaimTargetIsNotAcceptingClaims() : base("This character or group does not accept claims.") { }
-}
-
-public class InsufficientContactsException() : JoinRpgBaseException("Для отправки заявки необходимы контакты");
-
 public class RoomIsOccupiedException : JoinRpgProjectEntityException
 {
     public RoomIsOccupiedException(ProjectAccommodation entity) : base(entity, "Cannot peforrm this operation on occupied room.")
